Qt6に移行(今後Qt5ではビルド不可)
Squashed commit of the following:

commit 083c62e082bde5539c99e76cf96ba5df0710ed37
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 20:04:53 2022 +0900

    コードを整形

commit 4bedaaf996579f8042861fb2a2d4f20709ad84f3
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 20:04:10 2022 +0900

    ライブラリのインストール方法について追記

commit b7d65a5b99ffa1981a004324f79724b6d948b565
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 17:38:30 2022 +0900

    拡張カートリッジとコンボボックスの対応が間違っていたのを修正

commit 3610afd1ec979bfc68e4a99a401ddc8c13a13c17
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 14:10:17 2022 +0900

    MSYS2Privateを更新

commit c58984a0f8fb41ee74cf86e5625124dca39630e3
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 13:54:03 2022 +0900

    qt6を指定するように変更

commit 601dd3b122c227f53143a8c666de6596d0443a47
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 01:40:02 2022 +0900

    不要コードの整理

commit 3d8b6fc37491b636b5813c3c37af57f357d74dc7
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sun Feb 20 01:39:45 2022 +0900

    PULLモードでサウンド再生を実装

commit b8fe76c36f95f4272be05201ee1829ca64e595e3
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Sat Feb 19 23:11:53 2022 +0900

    ファイル選択ダイアログでフォルダしか選べなかったのを修正

commit 94c2893c16008f93ef513a7a16577a255a7a25eb
Merge: 7695a78 bb1c1a5
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Tue Feb 15 21:53:31 2022 +0900

    Merge branch 'master' into qt6

commit 7695a7805cb67f674524e0ece419cc57a6cdbc2b
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Mon Feb 14 22:12:22 2022 +0900

    とりあえず起動するようになった

commit fbd901cf8e4abef9158d97f0bd20f0a84d070cee
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Fri Feb 11 12:07:32 2022 +0900

    QtSingleApplicationをSingleApplicationに移行

commit 6afc802a94f1d37ef5d45fce5f4f16cd7cd1c0d2
Author: Tadahito Yao <eighttails@nifty.com>
Date:   Fri Feb 11 11:39:39 2022 +0900

    WIP

+++ win32/MSYS2Private/gimagereader/gimagereader.sh
@@ -41,14 +41,16 @@ if [ "$GIMAGEREADER_GIT" != "" ]; then
     # git pull
     GIMAGEREADER_BUILD_DIR=$GIMAGEREADER_SRC_DIR
 else
-    GIMAGEREADER_VERSION=master
-    if [ "$GIMAGEREADER_VERSION" == "master" ]; then
-        GIMAGEREADER_TAG=$GIMAGEREADER_VERSION
-        GIMAGEREADER_ARCHIVE=gImageReader-$GIMAGEREADER_TAG.tar.gz
-        rm $GIMAGEREADER_ARCHIVE 2> /dev/null
-    else
+    GIMAGEREADER_VERSION=5c04281d3b660308bcf3638c6e6a9da75675e432 #3.4.0+バグ修正
+    if [ "$GIMAGEREADER_VERSION" == "*.*" ]; then
         GIMAGEREADER_TAG=v$GIMAGEREADER_VERSION
         GIMAGEREADER_ARCHIVE=gImageReader-$GIMAGEREADER_TAG.tar.gz
+    else
+        GIMAGEREADER_TAG=$GIMAGEREADER_VERSION
+        GIMAGEREADER_ARCHIVE=gImageReader-$GIMAGEREADER_TAG.tar.gz
+        if [ "$GIMAGEREADER_VERSION" == "master" ]; then
+            rm $GIMAGEREADER_ARCHIVE 2> /dev/null
+        fi
     fi
     GIMAGEREADER_SRC_DIR=gImageReader-$GIMAGEREADER_VERSION
     GIMAGEREADER_BUILD_DIR=$GIMAGEREADER_SRC_DIR-$BIT
@@ -74,6 +76,7 @@ if [ "$GIMAGEREADER_DEBUG" != "" ]; then
 else
     BUILD_TYPE=Release
 fi
+
 CMAKE_PREFIX_PATH=$PREFIX/qt5-shared:$CMAKE_PREFIX_PATH \
 cmake .. \
 -G"MSYS Makefiles" \
